+++ packages/field/src/field/index.ts
@@ -1,4 +1,5 @@
 export * from "./field";
 export * from "./useClearInputAction";
+export * from "./useFieldError";
 export * from "./useFieldProps";
 export * from "./useRequiredProps";
