+++ packages/field/src/radio/RadioControl.tsx
@@ -1,11 +1,11 @@
-import { ValidatedFieldAtom } from "@form-atoms/field";
 import { atom } from "jotai";
 import { useMemo } from "react";
 import { RenderProp } from "react-render-prop-type";
 
-type AtomValue = ValidatedFieldAtom<boolean> | undefined;
+import { ValidatedFieldAtom } from "../field";
 
-const radioControlAtom = () => atom<AtomValue>(undefined);
+const radioControlAtom = () =>
+  atom<ValidatedFieldAtom<boolean> | undefined>(undefined);
 
 export type RadioControlAtom = ReturnType<typeof radioControlAtom>;
