+++ packages/field/src/radio/index.ts
@@ -1,1 +1,2 @@
 export * from "./Radio";
+export * from "./RadioControl";
