--- packages/flowbite/src/radio/index.ts
@@ -1,2 +1,0 @@
-export * from "./Radio";
-export * from "./RadioControl";
